+++ Package.swift
@@ -15,13 +15,13 @@ let package = Package(
         .package(url: "https://github.com/jpsim/Yams.git", from: "5.0.0"),
         .package(url: "https://github.com/yonaskolb/JSONUtilities.git", from: "4.2.0"),
         .package(url: "https://github.com/kylef/Spectre.git", from: "0.9.2"),
-        .package(url: "https://github.com/onevcat/Rainbow.git", from: "3.0.0"),
-        .package(url: "https://github.com/tuist/XcodeProj.git", from: "8.13.0"),
+        .package(url: "https://github.com/onevcat/Rainbow.git", from: "4.0.0"),
+        .package(url: "https://github.com/tuist/XcodeProj.git", exact: "8.13.0"),
         .package(url: "https://github.com/jakeheis/SwiftCLI.git", from: "6.0.3"),
         .package(url: "https://github.com/mxcl/Version", from: "2.0.0"),
         .package(url: "https://github.com/SwiftDocOrg/GraphViz.git", exact: "0.2.0"),
+        .package(url: "https://github.com/freddi-kit/ArtifactBundleGen", exact: "0.0.6")
     ],
-
     targets: [
         .executableTarget(name: "XcodeGen", dependencies: [
             "XcodeGenCLI",
@@ -42,6 +42,8 @@ let package = Package(
             "PathKit",
             "XcodeGenCore",
             "GraphViz",
+        ], resources: [
+            .copy("SettingPresets")
         ]),
         .target(name: "ProjectSpec", dependencies: [
             "JSONUtilities",